# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multiview Control API",
    "description": "REST + WebSocket + SSE management API for the Multiview engine.",
    "contact": {
      "name": "The Multiview Authors"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/outputs": {
      "get": {
        "tags": [
          "outputs"
        ],
        "summary": "`GET /api/v1/outputs` — list all outputs (role: read).",
        "operationId": "list_outputs",
        "responses": {
          "200": {
            "description": "All outputs, id-sorted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not authorized to read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/overlays": {
      "get": {
        "tags": [
          "overlays"
        ],
        "summary": "`GET /api/v1/overlays` — list all overlays (role: read).",
        "operationId": "list_overlays",
        "responses": {
          "200": {
            "description": "All overlays, id-sorted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not authorized to read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "`GET /api/v1/sources` — list all sources (role: read).",
        "operationId": "list_sources",
        "responses": {
          "200": {
            "description": "All sources, id-sorted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not authorized to read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "type": "object",
        "description": "An RFC 9457 problem-details document.\n\nSerialized as `application/problem+json`. The `status` field is authoritative\nand is also used as the HTTP status code by the [`IntoResponse`] impl. A\n`status` outside the valid `100..=599` range falls back to `500`.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "A human-readable explanation specific to this occurrence."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "A URI reference identifying this specific occurrence (e.g. the request\npath or an operation id)."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code, duplicated into the body per RFC 9457.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type."
          },
          "type": {
            "type": "string",
            "description": "A URI reference identifying the problem type (here a `/problems/<slug>`)."
          }
        }
      },
...
      "Resource": {
        "type": "object",
        "description": "A persisted management resource: a stable `id`, a display `name`, and the\nopaque `body` document (the config-as-code shape as canonical JSON).",
        "required": [
          "id",
          "name",
          "body"
        ],
        "properties": {
          "body": {
            "description": "The opaque resource document, as canonical JSON."
          },
          "id": {
            "type": "string",
            "description": "Stable resource id."
          },
          "name": {
            "type": "string",
            "description": "Human-friendly name."
          }
        }
      },
...
    }
  },
...
}